+++ cde/programs/dtinfo/DtMmdb/dti_cc/CC_Slist.C
@@ -46,7 +46,7 @@ template<class T>
 T *CC_TPtrSlist<T>::removeAt(size_t pos) {
 
   CC_TPtrSlistIterator<T> iter( *this );
-  for( int i = 0; i <= pos; i++ ) {
+  for( size_t i = 0; i <= pos; i++ ) {
     if ( !(++iter) ) {
       throw(CASTCCBEXCEPT ccBoundaryException(0,0,i));
     }
